Add `data` gulp task.
This adds a new Gulp task to automatically include `app/data` in
dev builds if it exists. This directory will contain sample data
files for use during development and outputs them to `build/data`.
Additionally, this integrates with browser-sync and will
automatically reload open browsers as testing datasets are
changed.

Change-Id: Ie1578682d11e36bd6142b9d326f50a547aa4613b