Add timeline view with Angular support.
This re-adds the Tempest timeline view as a set of Angular
directives. This includes related functionality, such as the Dstat
parser and some array utilities. The timeline view consists of
a timeline component, which includes the d3 chart, and a separate
details component, which shows additional information for tests
when selected in the timeline.

Change-Id: Ifaaeda91b0617e8cf7a60d30728005f5c8d00546