Add summary info to home page.
This patch adds basic summary information to the tempest dataset
summary on the home page. Additionally, it allows users to select
between all available datasets, if more than one is specified in
the configuration file. Failing tests are displayed directly on
the front page, and a future patch will allow direct navigation to
details of failing tests.

Co-Authored-By: Austin Clark <austin.clark@hpe.com>
Change-Id: Ic08a4dd86b36f513acf8e50c1fb7d7a102876e1c