Add key-listener to refactored timeline
This change adds a key-listener to navigate between test rectangles on
the timeline. When an out-of-view test is selected, the timeline will
refocus and bring it into view. Functionality for left and right arrow
keys is supported, up and down (to navigate between workers) key are
not supported, but may be added in a future patch.

Change-Id: I4a187a1c048d5e6552316d10db0c677432aa8f7e