Add test-details page
Patch for test-details page. Includes new page, directives,
controller, and route for navigation. The detail page URL follows
the same format as the timeline URL. Currently tabs are used to
display detail information, but this may be replaced in a future
revision. The page is navigated to from the timeline details
panel via a button in the bottom-right corner.

Change-Id: I003125a33f87aa7e4c8251756e1fe5f447abac56