Redesign the timeline details panel.
This redesigns the aspects of the timeline, particularly relating to
the details panel. First, selection mechanics are greatly simplified,
so items can be hovered while one is selected. Additionally, hovered
item names are now shown as part of the cursor. The details panel
now only shows details for selected items, and is moved to fill the
bottom of the screen; similarly, the timeline now fills the full width
of the screen.

Change-Id: I0acb88f898dcfd0a0327a38aa5eb995f118e1a1e