Add filter functionality to test-details pythonlogging
On tests with pythonlogging, the DEBUG headers can be quite long and
make finding useful information difficult. This patch adds a small,
four-button filter to the test-details page that allows a user to
select which levels of logging to be displayed (INFO, WARNING, ERROR,
DEBUG). This patch will likely serve as the base for additional search
functionality for test details (e.g. queries).

Change-Id: Ia4286861268de0e43313d3b8d7f966e94697a495