Show last two traceback lines in failure summary stats
Presently, the failure tracebacks displayed on the tempest summary
directive only show the last line of the failing test's traceback.
After analyzing some sample logs, it was determined that the last
two lines provide a more full picture of the failure, including
exception codes. This patchs adds code to the stackviz-export parser
to include the last two traceback lines for display on the summary
page.

Change-Id: Ib0bffd5b402093adbf1ff6618413df8e8a138670