Show warning for hidden tests on timeline.
Previously, tests hidden due to their small (< 2px) size would be
silently hidden from view, and navigating to a hidden test (via direct
link, search, or keyboard) would move the timeline to the correct
position, but would show only empty space, even though the details
panel would be updated. This adds a label and a warning message to the
details panel to inform users that the selected test is too small to
show in the view.

Additionally, when the selection is changed to a hidden item, any
previously-selected rect will now be properly restored to its original
color.

Change-Id: I3f6ee9313db64472ce8f6ce7d929c4d464f12ab5