Fix timeline worker sorting
In the nested 'data' array, nested lists of tests should be sorted by
their worker key. However, the 'd3.ascending' comparator sorts by
string rather than numerically, resulting in the array being out of
order with more than 10 workers. This caused item selection (but not
hovering) to break after the second row.

This patch replaces the built-in comparator ('d3.ascending') with a
comparator that uses parseInt() to make sure the nested lists are in
the correct order for fast lookups.

Change-Id: Idbd94369fb35f4cf03c39b0c9fa2cc07623885ab