Treat subunit as binary under python3
The subunit stream is a binary protocol and can't be encoded to utf8.
Unfortunately  under python3 the default behavior when reading from
stdin or opened files is to read them encoded as your platform dependent
encoding. For linux users this is typically utf8 and subunit isn't valid
utf8. Fix this by reading the subunit streams as binary under python3 as
we do on python2.

Change-Id: I1ddfe514c219c3bbfcd69f6808faa2331933acee