Fix dstat graph processing
This commit fixes dstat graph processing to draw a graph properly. There
are two problem mainly.

1. Reading header process is not correct
 It looks the dstat output format was changed to remove a row of the
header in the past. However, stackviz was not changed at that time. So,
this commit fixes to deal with it properly.

2. The year in its logs process is not correct.
 With the original process which sets 1900+year after utcParse() call,
it looks one hour difference occurs somehow. So, this commit sets the
year to the date string directly.

Change-Id: I52c321f4bd13475f3be2f1a9fbcd86a6ddd4e4be