Add enough testing on stackviz gate
While dropping py2 support, we dropped py3.5 support also
- Ic14a02c879bcef08885d831ebbe25083ec464a86

which broke the stable jobs using xenial node which does not
have py3.6.

This is fixing that and bring the py3.5 support back
- I0e6d57a49643d588b1329a0c8f1dd7afc71d5bdf

To avoid such issue in future, let's test the all users of
it which is nothing but integration jobs along with stable
branch jobs.

Change-Id: I7118fd5cfdeff7419f2477b64334767d13a41155