migrate to victoria testing runtime
As per victoria cycle testing runtime and community goal[1]
we need to migrate upstream CI/CD to Ubuntu Focal(20.04)
and Node.js 10.

Tox is not anymore pre-installed on the OpenDev CI images.
The file package.json needs tox, install it to fix the nodejs-job.

Install it via a pre-run playbook so that this is run in addition to all
of the playbooks of the parent.

This patch also update python version to py38 in package.json[2]
as py36 is not available on focal.

Note: Before this patch We used nodejs8-jobs from openstack-zuul-jobs
which consumes npm jobs from zuul-jobs. We would like to install tox
in the 'pre-run' phase in each npm job, so this patch also defines
npm job defination here.

Story: #2007865
Task: #40212

[1] https://governance.openstack.org/tc/goals/selected/victoria/migrate-ci-cd-jobs-to-ubuntu-focal.html
[2] https://opendev.org/openstack/stackviz/src/branch/master/package.json#L73

Co-Author: manchandavishal <manchandavishal143@gmail.com>

Change-Id: I116ea1609f88496c599a8285d33350104067c517