Move stable branch testing to oldest supported branch
Currently we test the stable branch with stable/train
which is in EM phase and also failing. Stackviz need to
test if supported branch works or not so let's test the
oldest stable supported branch. Considering stable/yoga
is going to EM/Unmaintained soon, I am adding stable/zed
job.

Change-Id: I5bf251f438317636aa0d2598718fe08a1aa97683